Remove type definitions from index
