+++ packages/wallet-specs/src/protocols/virtual-funding/protocol.ts
@@ -1,6 +1,6 @@
 import { assign } from 'xstate';
 
-import { Balance } from '../..';
+import { Balance } from '../../types';
 import { HubChoice } from '../../wire-protocol';
 import { Init as VirtualFundAsLeafArgs } from '../virtual-fund-as-leaf/protocol';
 import { store } from '../../temp-store';
